requirements.txt
asgiref==3.5.1
chardet==4.0.0
Django==3.2.13
et-xmlfile==1.1.0
numpy==1.21.6
openpyxl==3.0.9
pandas==1.3.5
plotly==5.7.0
python-dateutil==2.8.2
python-magic==0.4.25
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
typing_extensions==4.2.0
whitenoise==6.1.0
